+++ domain-model/src/main/java/uk/gov/hmcts/cmc/domain/utils/FeaturesUtils.java
@@ -11,4 +11,16 @@ public class FeaturesUtils {
     public static boolean hasMediationPilotFeature(Claim claim) {
         return claim.getFeatures() != null && claim.getFeatures().contains(ClaimFeatures.MEDIATION_PILOT.getValue());
     }
+
+    public static boolean isOnlineDQ(Claim claim) {
+        return claim.getFeatures() != null && claim.getFeatures().contains(ClaimFeatures.DQ_FLAG.getValue());
+    }
+
+    public static boolean isLegalAdvisorPilot(Claim claim) {
+        return claim.getFeatures() != null && claim.getFeatures().contains(ClaimFeatures.LA_PILOT_FLAG.getValue());
+    }
+
+    public static boolean isJudgePilot(Claim claim) {
+        return claim.getFeatures() != null && claim.getFeatures().contains(ClaimFeatures.JUDGE_PILOT_FLAG.getValue());
+    }
 }
