+++ src/main/java/uk/gov/hmcts/cmc/claimstore/events/claimantresponse/ClaimantResponseActionsHandler.java
@@ -6,12 +6,12 @@ import uk.gov.hmcts.cmc.claimstore.events.ccj.InterlocutoryJudgmentEvent;
 import uk.gov.hmcts.cmc.claimstore.services.notifications.NotificationToDefendantService;
 import uk.gov.hmcts.cmc.claimstore.services.staff.ClaimantRejectOrgPaymentPlanStaffNotificationService;
 import uk.gov.hmcts.cmc.claimstore.utils.ClaimantResponseHelper;
-import uk.gov.hmcts.cmc.claimstore.utils.DirectionsQuestionnaireUtils;
 import uk.gov.hmcts.cmc.domain.models.Claim;
 import uk.gov.hmcts.cmc.domain.models.claimantresponse.ClaimantResponse;
 import uk.gov.hmcts.cmc.domain.models.claimantresponse.ClaimantResponseType;
 import uk.gov.hmcts.cmc.domain.models.response.Response;
 import uk.gov.hmcts.cmc.domain.models.response.ResponseType;
+import uk.gov.hmcts.cmc.domain.utils.FeaturesUtils;
 
 import static uk.gov.hmcts.cmc.claimstore.utils.ClaimantResponseHelper.isOptedForMediation;
 import static uk.gov.hmcts.cmc.claimstore.utils.ResponseHelper.isOptedForMediation;
@@ -55,14 +55,14 @@ public class ClaimantResponseActionsHandler {
         ClaimantResponse claimantResponse = claim.getClaimantResponse().orElseThrow(IllegalStateException::new);
 
         return ClaimantResponseHelper.isIntentToProceed(claimantResponse)
-            && DirectionsQuestionnaireUtils.isOnlineDQ(claim);
+            && FeaturesUtils.isOnlineDQ(claim);
     }
 
     private boolean hasIntentionToProceedAndIsPaperDq(Claim claim) {
         ClaimantResponse claimantResponse = claim.getClaimantResponse().orElseThrow(IllegalStateException::new);
 
         return claimantResponse.getType() == REJECTION
-            && !DirectionsQuestionnaireUtils.isOnlineDQ(claim);
+            && !FeaturesUtils.isOnlineDQ(claim);
     }
 
     private boolean isFreeMediationConfirmed(Claim claim) {
